// frontend-hq-portal/src/types/smartCategory.ts
export interface SmartCategoryItemAssignment {
  itemId: number;
  itemCode: string;
  itemName: string;
  itemNameAlt?: string | null;
  displayIndex: number;
  enabled: boolean;
  modifiedDate: string;
  modifiedBy: string;
}
// ...
export interface SmartCategoryItemAssignmentEntry {
  itemId: number;
  displayIndex: number;
  enabled: boolean;
}